Deck Builder and Card Abstraction (#13)
* Deck Builder and Card Abstraction

Removed the deck building responsibility from the Deck, added a builder object with functions that define how to populate decks with cards.

Cards have been abstracted. There is now a Playing Card which contains most of the functionality of what Card was. Card is now a simple numeric value with the base methods.

This will allow for card games like Uno and Euchre where you want to build various decks of different card types.

* Cleanup from Code Review

Made some minor adjustments, then some major ones. Deck and Hand now
extend Array and their methods were moved to the prototype. This cleans
up all of the getCards() methods and proto methods are only defined once
so that reduces a bunch of overhead in dealing with a private array.

* Shifting to prototype

This really is the better way to do this, private properties just aren't worth it.